+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Next version
+
+- ApiControllerConcern : add `model_errors` in json response
+- `json_response` is now a `HashWithIndifferentAccess`
+
 ## 1.2.3
 
 - Add FCM